Modified webscraper, it now produces several csvs for each type of data ``` receiving_and_rushing.csv is not currently usable, maybe this is something for pandas to fix. I will look into this later.
I believe that the webscraper is now completed for the data that we will primarily be using.

I can attempt to comment some of the code to make it readable primarily for GUI implementation, I don't think we will use the scrape_all def or the main func, (python is still confusing me)
It is likely a better idea to modulize the code and call the scrape function as needed from elsewhere.